BREAKING: REMOVED CODE FOR OLD COMMANDS
