+++ Packages.props
@@ -3,7 +3,6 @@
         <PackageReference Update="Chronic.Core" Version="0.4.0" />
         <PackageReference Update="Cronos" Version="0.7.1" />
         <PackageReference Update="DSharpPlus" Version="4.3.*-*" />
-        <PackageReference Update="DSharpPlus.CommandsNext" Version="4.3.*-*" />
         <PackageReference Update="DSharpPlus.Interactivity" Version="4.3.*-*" />
         <PackageReference Update="DSharpPlus.SlashCommands" Version="4.3.*-*" />
         <PackageReference Update="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="6.0.6" />
