+++ src/HonzaBotner.Discord.Services/Commands/MemberCommands.cs
@@ -1,15 +1,11 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
-using DSharpPlus.CommandsNext;
-using DSharpPlus.CommandsNext.Attributes;
+using DSharpPlus;
 using DSharpPlus.Entities;
-using DSharpPlus.EventArgs;
 using DSharpPlus.Exceptions;
-using DSharpPlus.Interactivity;
-using DSharpPlus.Interactivity.Extensions;
+using DSharpPlus.SlashCommands;
 using HonzaBotner.Database;
-using HonzaBotner.Discord.Services.Attributes;
 using HonzaBotner.Discord.Services.Options;
 using HonzaBotner.Services.Contract;
 using Microsoft.EntityFrameworkCore;
@@ -18,19 +14,16 @@ using Microsoft.Extensions.Options;
 
 namespace HonzaBotner.Discord.Services.Commands;
 
-[Group("member")]
-[Description("Commands to interact with members.")]
-[ModuleLifespan(ModuleLifespan.Transient)]
-[RequireGuild]
-public class MemberCommands : BaseCommandModule
+[SlashCommandGroup("member", "member commands")]
+[GuildOnly]
+[SlashCommandPermissions(Permissions.ModerateMembers)]
+public class MemberCommands : ApplicationCommandModule
 {
-    [Group("info")]
-    [Aliases("about", "whois")]
-    [Description("Provides info about a member.")]
-    [RequireMod]
-    [ModuleLifespan(ModuleLifespan.Transient)]
-    public class MemberCommandsInfo : BaseCommandModule
+
+    [SlashCommandGroup("info", "Provides info about a member.")]
+    public class MemberCommandsInfo : ApplicationCommandModule
     {
+
         private readonly HonzaBotnerDbContext _dbContext;
         private readonly IHashService _hashService;
         private readonly ILogger<MemberCommandsInfo> _logger;
@@ -38,82 +31,58 @@ public class MemberCommands : BaseCommandModule
         public MemberCommandsInfo(
             HonzaBotnerDbContext dbContext,
             IHashService hashService,
-            ILogger<MemberCommandsInfo> logger
-        )
+            ILogger<MemberCommandsInfo> logger)
         {
             _dbContext = dbContext;
             _hashService = hashService;
             _logger = logger;
         }
 
-        [GroupCommand]
-        [Command("discord")]
-        [Description("Provides info about a member by Discord account.")]
-        public async Task DiscordMemberInfo(
-            CommandContext ctx,
-            [Description("Member to show info about.")]
-            DiscordMember member
-        )
+        [SlashCommand("discord", "Provides info about a member by Discord account.")]
+        public async Task DiscordMemberInfoCommandAsync(
+            InteractionContext ctx,
+            [Option("member","Who is the target?")] DiscordUser user)
         {
             Verification? databaseRecord = await _dbContext.Verifications
-                .FirstOrDefaultAsync(v => v.UserId == member.Id);
+                .FirstOrDefaultAsync(v => v.UserId == user.Id);
             await AnnounceMemberInfoAsync(ctx, databaseRecord);
         }
 
-        [Command("ctu")]
-        [Aliases("cvut")]
-        [Description("Provides info about a member by CTU username.")]
-        public async Task CtuMemberInfo(
-            CommandContext ctx,
-            [Description("CTU username.")] string cvutUsername
-        )
+        [SlashCommand("ctu", "Provides info about a member by CTU username.")]
+        public async Task CtuMemberInfoCommandAsync(
+            InteractionContext ctx,
+            [Option("username", "Who is the target?")] string cvutUsername)
         {
             string authId = _hashService.Hash(cvutUsername);
-            Verification? databaseRecord =
-                await _dbContext.Verifications.FirstOrDefaultAsync(v => v.AuthId == authId);
+            Verification? databaseRecord = await _dbContext.Verifications
+                .FirstOrDefaultAsync(v => v.AuthId == authId);
             await AnnounceMemberInfoAsync(ctx, databaseRecord);
         }
 
-        private async Task AnnounceMemberInfoAsync(CommandContext ctx, Verification? databaseRecord)
+        private async Task AnnounceMemberInfoAsync(InteractionContext ctx, Verification? databaseRecord)
         {
             if (databaseRecord == null)
             {
-                await ctx.Channel.SendMessageAsync("No member record.");
+                await ctx.CreateResponseAsync("No member record for provided name");
                 return;
             }
 
             try
             {
-                DiscordMember? member = await ctx.Guild.GetMemberAsync(databaseRecord.UserId);
-                await member.SendMessageAsync(
-                    $"Member {ctx.Member.DisplayName} requested information about your account on the FIT CTU Discord server."
-                );
-            }
-            catch (UnauthorizedException e)
-            {
-                _logger.LogWarning(
-                    e,
-                    "Couldn't send message to user {UserId} because" +
-                    " the member is no longer in the guild" +
-                    " or the member has Allow DM from server members off",
-                    databaseRecord.UserId
-                );
+                DiscordMember member = await ctx.Guild.GetMemberAsync(databaseRecord.UserId);
+                await member.SendMessageAsync($"Member {ctx.Member.DisplayName} requested information about your account on the FIT CTU Discord server.");
             }
             catch (Exception e)
             {
                 _logger.LogWarning(e, "Couldn't get member {MemberId}", databaseRecord.UserId);
             }
 
-            await ctx.Channel.SendMessageAsync(databaseRecord.ToString());
+            await ctx.CreateResponseAsync(databaseRecord.ToString());
         }
     }
 
-    [Group("delete")]
-    [Aliases("erase", "remove")]
-    [Description("Erases database record of the member.")]
-    [RequireMod]
-    [ModuleLifespan(ModuleLifespan.Transient)]
-    public class MemberCommandsDelete : BaseCommandModule
+    [SlashCommandGroup("delete", "Erases database record of the member.")]
+    public class MemberCommandsDelete : ApplicationCommandModule
     {
         private readonly HonzaBotnerDbContext _dbContext;
         private readonly IHashService _hashService;
@@ -124,8 +93,7 @@ public class MemberCommands : BaseCommandModule
             HonzaBotnerDbContext dbContext,
             IHashService hashService,
             ILogger<MemberCommandsDelete> logger,
-            IGuildProvider guildProvider
-        )
+            IGuildProvider guildProvider)
         {
             _dbContext = dbContext;
             _hashService = hashService;
@@ -133,57 +101,35 @@ public class MemberCommands : BaseCommandModule
             _guildProvider = guildProvider;
         }
 
-        [GroupCommand]
-        [Command("discord")]
-        [Description("Erases database record of the member by Discord account.")]
-        public async Task DiscordMemberErase(
-            CommandContext ctx,
-            [Description("Discord member to erase.")]
-            DiscordMember member
-        )
+        [SlashCommand("discord", "Erases database record of the member by Discord account.")]
+        public async Task DiscordMemberEraseCommandAsync(
+            InteractionContext ctx,
+            [Option("member", "Who is the target?")] DiscordUser user)
         {
-            Verification? databaseRecord =
-                await _dbContext.Verifications.FirstOrDefaultAsync(v => v.UserId == member.Id);
-            await EraseMemberAsync(ctx, databaseRecord, member.DisplayName);
+            Verification? databaseRecord = await _dbContext.Verifications
+                .FirstOrDefaultAsync(v => v.UserId == user.Id);
+            await EraseMemberAsync(ctx, databaseRecord);
         }
 
-        [Command("ctu")]
-        [Aliases("cvut")]
-        [Description("Erases database record of the member by CTU username.")]
-        public async Task CtuMemberErase(
-            CommandContext ctx,
-            [Description("CVUT username of member to erase.")]
-            string cvutUsername
-        )
+        [SlashCommand("ctu", "Erases database record of the member by CTU username.")]
+        public async Task CtuMemberEraseCommandAsync(
+            InteractionContext ctx,
+            [Option("username", "Who is the target?")] string cvutUsername)
         {
             string authId = _hashService.Hash(cvutUsername);
-            Verification? databaseRecord =
-                await _dbContext.Verifications.FirstOrDefaultAsync(v => v.AuthId == authId);
-            await EraseMemberAsync(ctx, databaseRecord, cvutUsername);
+            Verification? databaseRecord = await _dbContext.Verifications
+                .FirstOrDefaultAsync(v => v.AuthId == authId);
+            await EraseMemberAsync(ctx, databaseRecord);
         }
 
-        private async Task EraseMemberAsync(CommandContext ctx, Verification? databaseRecord, string userName)
+        private async Task EraseMemberAsync(InteractionContext ctx, Verification? databaseRecord)
         {
+            await ctx.DeferAsync();
             if (databaseRecord == null)
             {
-                await ctx.Channel.SendMessageAsync("No member record to erase.");
+                await ctx.CreateResponseAsync("No member record to erase.");
                 return;
             }
-
-            DiscordEmoji emoji = DiscordEmoji.FromName(ctx.Client, ":ok_hand:");
-            DiscordMessage reactMessage =
-                await ctx.Channel.SendMessageAsync($"To approve erase of `{userName}`, react with {emoji}");
-            await reactMessage.CreateReactionAsync(emoji);
-            InteractivityResult<MessageReactionAddEventArgs> result =
-                await reactMessage.WaitForReactionAsync(ctx.Member, emoji, TimeSpan.FromSeconds(30));
-
-            if (result.TimedOut)
-            {
-                await ctx.Channel.SendMessageAsync("Member hasn't been erased due to approval timeout.");
-                return;
-            }
-
-            await ctx.TriggerTypingAsync();
 
             try
             {
@@ -192,11 +138,13 @@ public class MemberCommands : BaseCommandModule
                 DiscordGuild guild = await _guildProvider.GetCurrentGuildAsync();
                 DiscordMember member = await guild.GetMemberAsync(databaseRecord.UserId);
                 await member.RemoveAsync("This user was kicked due to being purged from the database.");
-                await ctx.Channel.SendMessageAsync("Member has been erased.");
+                await ctx.CreateResponseAsync("Member has been erased.");
+                return;
             }
-            catch (UnauthorizedException)
+            catch (UnauthorizedException e)
             {
-                await ctx.RespondAsync("User was purged but not kicked due to insufficient permissions\n" +
+                _logger.LogWarning(e, "Erasing of member failed due to lack of permissions");
+                await ctx.CreateResponseAsync("User was purged but not kicked due to insufficient permissions\n" +
                                        "Please remove verified role manually to prevent unexpected behaviour.");
             }
             catch (Exception e)
@@ -204,83 +152,60 @@ public class MemberCommands : BaseCommandModule
                 await ctx.Channel.SendMessageAsync("Member erase failed.");
                 _logger.LogWarning(e, "Member erase failed");
             }
+
+            await ctx.CreateResponseAsync("Failed to erase member. More information in the log.");
         }
     }
 
-    [Group("count")]
-    [Description("Counts members by provided roles.")]
-    [RequireAllowlist(AllowlistsTypes.MemberCount)]
-    [ModuleLifespan(ModuleLifespan.Transient)]
-    public class MemberRoleCount : BaseCommandModule
+    [SlashCommandGroup("count", "Counts members by provided roles.")]
+    public class MemberCommandsCount : ApplicationCommandModule
     {
         private readonly CommonCommandOptions _commonCommandOptions;
 
-        public MemberRoleCount(IOptions<CommonCommandOptions> commonCommandOptions)
+        public MemberCommandsCount(IOptions<CommonCommandOptions> commonCommandOptions)
         {
             _commonCommandOptions = commonCommandOptions.Value;
         }
 
-        [Command("all")]
-        [Description("Counts all members and all authenticated members.")]
-        public async Task CountAll(CommandContext ctx)
+        [SlashCommand("all", "Counts all members and all authenticated members.")]
+        public async Task CountAllCommandAsync(InteractionContext ctx)
         {
             DiscordGuild guild = ctx.Guild;
             DiscordRole authenticatedRole = guild.GetRole(_commonCommandOptions.AuthenticatedRoleId);
 
-            int authenticatedCount = 0;
-            foreach ((_, DiscordMember member) in guild.Members)
-            {
-                if (member.Roles.Contains(authenticatedRole))
-                {
-                    authenticatedCount++;
-                }
-            }
-
-            await ctx.Channel.SendMessageAsync(
-                $"Authenticated: {authenticatedCount}, All: {ctx.Guild.Members.Count.ToString()}");
+            int authenticatedCount = guild.Members.Count(member => member.Value.Roles.Contains(authenticatedRole));
+            await ctx.CreateResponseAsync($"Authenticated: {authenticatedCount}, All: {ctx.Guild.Members.Count.ToString()}");
         }
 
-        [GroupCommand]
-        [Command("roleOr")]
-        [Aliases("or")]
-        [Description("Counts all members which have AT LEAST ONE of the provided roles.")]
-        public async Task CountRoleOr(
-            CommandContext ctx,
-            [Description("Roles to check.")] params DiscordRole[] roles
-        )
+        [SlashCommand("here", "Counts all members who can see this channel")]
+        public async Task CountHereCommandAsync(
+            InteractionContext ctx,
+            [Option("ephemeral", "Hide response? Default false")] bool ephemeral = false)
+        {
+            await ctx.CreateResponseAsync($"Members in this channel: {ctx.Channel.Users.Count}", ephemeral);
+        }
+
+        [SlashCommand("role", "Counts members based on provided roles")]
+        public async Task CountRoleCommandAsync(
+            InteractionContext ctx,
+            [Option("roles", "Roles to look for - write as a mention")] string roles,
+            [Choice("and", "and")]
+            [Choice("or", "or")]
+            [Option("search-type", "Look for members with all those roles/some. Default: and")] string type = "and")
         {
             int count = 0;
             DiscordGuild guild = ctx.Guild;
 
-            foreach ((_, DiscordMember member) in guild.Members)
-            {
-                if (roles.Any(role => member.Roles.Contains(role)))
-                {
-                    count++;
-                }
-            }
-
-            await ctx.Channel.SendMessageAsync(count.ToString());
-        }
-
-        [Command("roleAnd")]
-        [Aliases("and")]
-        [Description("Counts all members which have ALL of the provided roles.")]
-        public async Task CountRoleAnd(
-            CommandContext ctx,
-            [Description("Roles to check.")] params DiscordRole[] roles
-        )
-        {
-            int count = 0;
-            DiscordGuild guild = ctx.Guild;
+            await ctx.DeferAsync();
 
             foreach ((_, DiscordMember member) in guild.Members)
             {
-                bool hasAllRoles = roles.All(role => member.Roles.Contains(role));
-
-                if (hasAllRoles)
+                switch (type)
                 {
-                    count++;
+                    case "or" when ctx.ResolvedRoleMentions.Any(role => member.Roles.Contains(role)):
+                    case "and" when ctx.ResolvedRoleMentions.All(role => member.Roles.Contains(role)):
+                        count++;
+                        break;
                 }
             }
