+++ src/HonzaBotner.Discord.Services/Commands/Polls/PollException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace HonzaBotner.Discord.Services.SCommands.Polls;
+namespace HonzaBotner.Discord.Services.Commands.Polls;
 
 public class PollException : Exception
 {
