--- src/HonzaBotner.Discord.Services/Commands/TestCommands.cs
@@ -1,19 +1,0 @@
-﻿using System.Threading.Tasks;
-using DSharpPlus.CommandsNext;
-using DSharpPlus.CommandsNext.Attributes;
-using HonzaBotner.Discord.Services.Attributes;
-
-namespace HonzaBotner.Discord.Services.Commands;
-
-[Group("test")]
-[Description("Testing commands")]
-[RequireMod]
-[RequireGuild]
-public class TestCommands : BaseCommandModule
-{
-    [Command("boostCount")]
-    public async Task BoostCount(CommandContext ctx)
-    {
-        await ctx.Channel.SendMessageAsync(ctx.Guild.PremiumSubscriptionCount.ToString());
-    }
-}
