--- src/HonzaBotner.Discord/ICommandConfigurator.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using DSharpPlus.CommandsNext;
-
-namespace HonzaBotner.Discord;
-
-public class CommandConfigurator
-{
-    public CommandConfigurator(Action<CommandsNextExtension> config) => Config = config;
-
-    public Action<CommandsNextExtension> Config { get; }
-}
