+++ src/HonzaBotner/appsettings.BotDev.json
@@ -7,10 +7,7 @@
         }
     },
     "Discord": {
-        "GuildId": 750055928669405258,
-        "Prefixes": [
-            "::", "Botner::", "cz.cvut.fit.botner."
-        ]
+        "GuildId": 750055928669405258
     },
     "CVUT": {
         "AppBaseUrl": "https://fit-cvut-discord-test.herokuapp.com"
@@ -30,17 +27,9 @@
         "StandUpRoleId": 0,
         "StandUpChannelId": 0
     },
-    "CommandAllowlistsOptions": {
-        "MemberCount": [
-            809436711507066920
-        ]
-    },
     "CustomVoiceOptions": {
         "ClickChannelId": 810277031089930251,
-        "RemoveAfterCommandInSeconds": 30,
-        "CommandChannelsIds": [
-            750108543125946448
-        ]
+        "RemoveAfterCommandInSeconds": 30
     },
     "PinOptions": {
         "Threshold": 10,
