+++ src/HonzaBotner/appsettings.CvutFit.json
@@ -11,11 +11,6 @@
     },
     "Discord": {
         "GuildId": 366970031445377024,
-        "Prefixes": [
-            "::",
-            "Botner::",
-            "cz.cvut.fit.botner."
-        ],
         "LogChannelId": 888087171712565289
     },
     "CommonCommandOptions": {
@@ -37,18 +32,9 @@
         "StandUpRoleId": 972019490818768927,
         "StandUpChannelId": 972019138446897162
     },
-    "CommandAllowlistsOptions": {
-        "MemberCount": [
-            681642388300890140
-        ]
-    },
     "CustomVoiceOptions": {
         "ClickChannelId": 811176917146271764,
-        "RemoveAfterCommandInSeconds": 30,
-        "CommandChannelsIds": [
-            814601150648942603,
-            681560465066688523
-        ]
+        "RemoveAfterCommandInSeconds": 30
     },
     "PinOptions": {
         "Threshold": 10,
