+++ src/HonzaBotner/appsettings.Development.json
@@ -7,21 +7,16 @@
         }
     },
     "Discord": {
-        "GuildId": 750055928669405258,
-        "Prefixes": [
-            "::",
-            "Botner::",
-            "cz.cvut.fit.botner."
-        ],
-        "LogChannelId": 888086696023973938
+        "GuildId": 887815108720467998,
+        "LogChannelId": 968247366849282062
     },
     "CVUT": {
         "AppBaseUrl": "https://localhost:5001"
     },
     "CommonCommandOptions": {
-        "ModRoleId": 811385023705645086,
-        "BotRoleId": 750102300453765202,
-        "AuthenticatedRoleId": 809436691470614539,
+        "ModRoleId": 887978558759264296,
+        "BotRoleId": 887978852930953246,
+        "AuthenticatedRoleId": 887977978712174622,
         "MuteRoleId": 0,
         "HugEmoteName": "",
         "BoosterRoleId": 820565262095876158,
@@ -33,27 +28,16 @@
         "StandUpRoleId": 972123386094440498,
         "StandUpChannelId": 750108543125946448
     },
-    "CommandAllowlistsOptions": {
-        "MemberCount": [
-            809436711507066920
-        ]
-    },
     "CustomVoiceOptions": {
-        "ClickChannelId": 810277031089930251,
-        "RemoveAfterCommandInSeconds": 30,
-        "CommandChannelsIds": [
-            750108543125946448
-        ]
+        "ClickChannelId": 887815109165076500,
+        "RemoveAfterCommandInSeconds": 30
     },
     "PinOptions": {
-        "Threshold": 5,
+        "Threshold": 1,
         "TemporaryPinName": ":pushpin:",
         "PermanentPinName": ":round_pushpin:",
         "LockEmojiName": ":lock:",
-        "RoleToWeightMapping": {
-            "809436691470614539": 5,
-            "809436711507066920": 3
-        }
+        "RoleToWeightMapping": {}
     },
     "ReminderOptions": {
         "CancelEmojiName": "\uD83D\uDD07",
